Breakdown time in years

+++ src/sql/packaging/schemaPackaging.sql
@@ -2,6 +2,6 @@ CREATE TABLE packaging (
     id SERIAL UNIQUE,
     name VARCHAR(128) NOT NULL PRIMARY KEY,
     footprint REAL NOT NULL,
-    breakdown_time INT NOT NULL,
+    breakdown_time REAL NOT NULL,
     reusability INT NOT NULL
 );
